scripts/requirements.txt
python-dotenv>=1.0.0
psycopg[binary]>=3.1.0
clickhouse-connect>=0.7.0
pyyaml>=6.0
